backend/requirements.txt
flask
opencv-python
numpy
ultralytics
rembg
pillow